+++ src/main/java/com/adore/convenientcrafting/ConvenientCrafting.java
@@ -3,6 +3,7 @@ package com.adore.convenientcrafting;
 import org.slf4j.Logger;
 
 import com.adore.convenientcrafting.config.Config;
+import com.adore.convenientcrafting.event.BagPickupEvents;
 import com.adore.convenientcrafting.network.CraftRecipePacket;
 import com.adore.convenientcrafting.network.NestedCraftingMissingMaterialsPacket;
 import com.adore.convenientcrafting.network.RecipeUnlockSyncPacket;
@@ -56,6 +57,7 @@ public class ConvenientCrafting {
         ModCreativeModeTabs.register(modEventBus);
 
         NeoForge.EVENT_BUS.register(this);
+        NeoForge.EVENT_BUS.register(new BagPickupEvents());
         NeoForge.EVENT_BUS.register(new RecipeUnlocks());
 
         modContainer.registerConfig(ModConfig.Type.COMMON, Config.SPEC);